# openapi/IDIS_OpenAPI_v6_3.yaml
openapi: 3.0.3
info:
  title: IDIS API (VC Edition) - v6.3
  version: "6.3"
  description: >
    Enterprise-grade API contract for IDIS (Institutional Deal Intelligence System), derived from IDIS v6.3.
    This API enables deal intake, document ingestion, claim registry access, Sanad chain retrieval,
    deterministic calculation runs, LangGraph debate runs, human verification gates, overrides,
    deliverable generation, audit logging, and integrations.
servers:
  - url: https://api.idis.example.com
    description: Production
  - url: https://api-staging.idis.example.com
    description: Staging
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/deals/{dealId}/assignments:
    post:
      tags: [AccessAdmin]
      summary: Assign an actor to a deal (ADMIN only)
      operationId: createDealAssignment
      responses:
        "201": { description: Assignment created }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [actor_id]
              properties:
                actor_id: { type: string, minLength: 1 }
  /v1/deals/{dealId}/assignments/{actorId}:
    delete:
      tags: [AccessAdmin]
      summary: Remove a direct deal assignment (ADMIN only)
      operationId: deleteDealAssignment
      responses:
        "204": { description: Assignment removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/groups:
    post:
      tags: [AccessAdmin]
      summary: Create a group (ADMIN only)
      operationId: createGroup
      responses:
        "201": { description: Group created }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string }
  /v1/groups/{groupId}/members:
    post:
      tags: [AccessAdmin]
      summary: Add an actor to a group (ADMIN only)
      operationId: addGroupMember
      responses:
        "201": { description: Member added }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [actor_id]
              properties:
                actor_id: { type: string, minLength: 1 }
  /v1/groups/{groupId}/members/{actorId}:
    delete:
      tags: [AccessAdmin]
      summary: Remove an actor from a group (ADMIN only)
      operationId: removeGroupMember
      responses:
        "204": { description: Member removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/deals/{dealId}/group-assignments:
    post:
      tags: [AccessAdmin]
      summary: Assign a group to a deal (ADMIN only)
      operationId: assignGroupToDeal
      responses:
        "201": { description: Group assigned }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [group_id]
              properties:
                group_id: { type: string, minLength: 1 }
  /v1/deals/{dealId}/group-assignments/{groupId}:
    delete:
      tags: [AccessAdmin]
      summary: Unassign a group from a deal (ADMIN only)
      operationId: unassignGroupFromDeal
      responses:
        "204": { description: Group unassigned }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/break-glass/grants:
    post:
      tags: [BreakGlass]
      summary: Issue a single-use break-glass grant for the requesting admin (ADMIN only)
      operationId: createBreakGlassGrant
      responses:
        "201":
          description: Grant issued; the token is returned once and never logged or audited raw
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [grant_id, token, expires_at, deal_id]
                properties:
                  grant_id: { type: string, format: uuid }
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
                  deal_id: { type: string, format: uuid }
        "400": { description: Invalid justification }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Forbidden }
        "404": { $ref: "#/components/responses/NotFound" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [deal_id, justification]
              properties:
                deal_id: { type: string, format: uuid }
                justification: { type: string, minLength: 1 }
                duration_seconds:
                  type: integer
                  description: Requested validity window; the core clamps to [60, 3600] seconds
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-IDIS-API-Key
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "bad_request" }
        message: { type: string, example: "Invalid field: deal.name" }
        details:
          type: object
          additionalProperties: true
        request_id: { type: string, example: "req_123" }
      example:
        code: "not_found"
        message: "Deal a1b2c3d4-e5f6-7890-abcd-ef1234567890 not found"
        details:
          resource_type: "deal"
          resource_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        request_id: "req_8f3a2b1c-0000-0000-0000-000000000001"
...